requirements.txt
requests>=2.31,<3.0
streamlit>=1.30,<2.0
pandas>=2.0,<3.0
scikit-learn>=1.3,<2.0
python-dotenv>=1.0,<2.0
plotly>=5.18,<6.0
numpy>=1.24,<3.0
fastapi>=0.109,<1.0
uvicorn[standard]>=0.27,<1.0
